requirements.txt
asgiref==3.7.2
asyncio==3.4.3
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
Faker==24.11.0
greenlet==3.0.3
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
schedule==1.2.2
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
websockets==12.0
